relocate Schools::RegisterMentor::Completion to a more general place
