+++ spec/services/schools/register_mentor_spec.rb
@@ -1,4 +1,4 @@
-describe Schools::RegisterMentor::Completion do
+describe Schools::RegisterMentor do
   let(:first_name) { "Dusty" }
   let(:last_name) { "Rhodes" }
   let(:trn) { "3002586" }
@@ -18,14 +18,14 @@ describe Schools::RegisterMentor::Completion do
     let(:mentor_at_school_period) { MentorAtSchoolPeriod.first }
 
     it 'creates a new Teacher record' do
-      expect { service.call }.to change(Teacher, :count).from(0).to(1)
+      expect { service.register_mentor! }.to change(Teacher, :count).from(0).to(1)
       expect(teacher.first_name).to eq(first_name)
       expect(teacher.last_name).to eq(last_name)
       expect(teacher.trn).to eq(trn)
     end
 
     it 'creates an associated MentorATSchoolPeriod record' do
-      expect { service.call }.to change(MentorAtSchoolPeriod, :count).from(0).to(1)
+      expect { service.register_mentor! }.to change(MentorAtSchoolPeriod, :count).from(0).to(1)
       expect(mentor_at_school_period.teacher_id).to eq(teacher.id)
       expect(mentor_at_school_period.started_on).to eq(started_on)
     end
@@ -39,7 +39,7 @@ describe Schools::RegisterMentor::Completion do
       end
 
       it "current date is assigned" do
-        service.call
+        service.register_mentor!
 
         expect(mentor_at_school_period.started_on).to eq(Date.current)
       end
